Chore: move utils/firebase-config.js to config/firebase.js
